cpuidle: Fix incorrect optimization. commit 672917dcc781ead7652a8b11b1fba14e38ac15b8 added an optimization, where the analysis on the past idle period moved from the end of idle, to the beginning of the new idle.
Unfortunately, this optimization had a bug where it zeroed one key variable
for new use, that is needed for the analysis. The fix is simple,
zero the variable after doing the work from the previous idle.

During the audit of the code that found this issue, another issue was
also found; the ->measured_us data structure member is never set,
a local variable is always used instead.

Signed-off-by: Arjan van de Ven <arjan@linux.intel.com>