timer: provide microaccounting for iowait in addition to idle time + nohz: Track last do_timer() cpu. Today, the kernel has microaccounting for the CPU idle time, as used by cpufreq governors and other places.
For a cpufreq governor I'm working on I would also like to get access
to microaccounting of iowait time.

This patch adds the iowait micro-accounting to the existing microaccounting
code. In addition, I've improved the accuracy of the data by updating
the totals at the time of asking for the total, rather than just reporting
the total at the last schedule point.

Signed-off-by: Arjan van de Ven <arjan@...>

The previous patch which limits the sleep time to the maximum
deferment time of the time keeping clocksource has some limitations on
SMP machines: if all CPUs are idle then for all CPUs the maximum sleep
time is limited.

Solve this by keeping track of which cpu had the do_timer() duty
assigned last and limit the sleep time only for this cpu.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
LKML-Reference: <new-submission>
Cc: Jon Hunter <jon-hunter@ti.com>
Cc: John Stultz <johnstul@us.ibm.com>